Issue #39 Change order type of annotations from int do touble

+++ src/main/java/org/jboss/qa/jenkins/test/executor/phase/download/Download.java
@@ -41,5 +41,5 @@ public @interface Download {
 	UnPack unpack() default @UnPack;
 
 	@Order
-	int order() default 0;
+	double order() default 0;
 }
